Display logo & suppress photo notice on logo
Had a chat about this last night on Indieweb chat. Just finished updating my site as a test-candidate.

As a user who is deliberately opting not to display a photograph of myself, who would rather be represented by a graphic than a photograph, in the absence of avatar support. I would like to see my logo displayed via indiewebify.me, and have the helpful message to include a photo suppressed, if I supply a logo, but not a photo; so that I'm not bothered by the message and do not feel friction with the choice to represent myself in this way.